ComE/requirements.txt
certifi==2019.11.28
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
dorthrithil-networkx==1.11
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.3
oldest-supported-numpy==0.2
psutil==5.6.7
pyparsing==2.4.6
python-dateutil==2.8.1
scikit-learn==0.22.1
scipy==1.4.1
SDG-app==0.0.0
six==1.14.0
sklearn==0.0
wincertstore==0.2
